external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/ternary_int_float.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:4: ':' :  wrong operand types  no operation ':' exists that takes a left-hand operand of type 'const mediump float' and a right operand of type 'const mediump int' (or there is no acceptable conversion)
ERROR: 0:4: '=' :  cannot convert from 'const mediump int' to 'float'

implicit cast of int to float in ternary expression should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast of int to float in ternary expression should fail
test vertex shader
default fragment shaderPASS implicit cast of int to float in ternary expression should fail
PASS successfullyParsed is true

TEST COMPLETE
